flask app folder structure organized by feature and replace hardcoded local endpoint with relative endpoints
