+++ server/app/mrds/__init__.py
@@ -3,4 +3,4 @@ from flask import Blueprint
 # get the blueprint instance
 mrds_bp = Blueprint('mrds', __name__)
 
-from . import routes
+from app.mrds import routes
