+++ server/app/mrds/routes.py
@@ -1,4 +1,4 @@
-from flask import Flask, jsonify, request, send_file, Blueprint
+from flask import jsonify, request, send_file
 from matplotlib import pyplot as plt
 from scipy.io import loadmat
 import os
