+++ server/app/viewer/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+viewer_bp = Blueprint("viewer", __name__)
+
+from . import routes
