--- server/app/visualize/__init__.py
@@ -1,2 +1,0 @@
-from flask import Blueprint
-
