--- server/app/visualize/magnets/clinical_processing.py
@@ -1,46 +1,0 @@
-"""
-Module: clinical_processing.py
-
-Description:
-This module handles the processing of HP MRI data from clinical-grade MRI machines.
-It includes functionality for adjusting image processing techniques to meet the needs
-of clinical data formats and imaging characteristics.
-
-Functions:
-- process_proton_picture(slider_value): Retrieves and processes proton images for clinical magnets based on user input.
-- process_hpmri_data(epsi_value, threshold): Filters and processes HP MRI data for clinical use, applying specified thresholds.
-- read_epsi_plot(epsi_value, threshold): Customizes EPSI plot processing to suit the specifics of clinical MRI data.
-
-Author:
-Benjamin Yoon
-
-Date:
-2024
-
-Version:
-1.2.2
-"""
-
-# Import statements
-import numpy as np
-from flask import jsonify
-import os
-
-
-# Example of function definition
-def process_proton_picture(slider_value, data):
-    """
-    Retrieves a proton image from clinical MRI data based on a slider value.
-    Adjusts image processing to fit clinical standards and data specifications.
-
-    Args:
-        slider_value (int): The index to determine which image to load.
-        data (dict): Additional data necessary for processing the image, such as contrast settings or other parameters.
-
-    Returns:
-        Flask Response: Image file as PNG or an error message in JSON format.
-    """
-    pass
-
-
-# Additional functions would follow with similar structured comments...
